refactor: move CurrentUser class to model.dto package and update imports
